reorganisation of ports, inplace and not_inplace module
